# api/openapi.yaml
openapi: 3.0.3
info:
  title: SmartLogix API
  version: '1.0.0'
  description: API unificada (BFF) para SmartLogix — documentación principal para la entrega.
servers:
  - url: http://localhost:4000
    description: BFF local
paths:
...
  /api/payment/create-intent:
    post:
      summary: Crear intención de pago
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentIntentRequest'
      responses: { '200': { description: 'Intent creado' } }
  /api/payment/{paymentId}/status:
    get:
      summary: Obtener estado de pago
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      responses: { '200': { description: 'Estado' } }
  /api/payment/{paymentId}/refund:
    post:
      summary: Reembolso
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      responses: { '200': { description: 'Reembolso procesado' } }
...
components:
  schemas:
...
    PaymentIntentRequest:
      type: object
      properties:
        amount: { type: number }
        currency: { type: string }
        orderId: { type: string }